requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.36
pydantic==2.9.2
Pillow==10.4.0
python-dotenv==1.0.1
streamlit==1.38.0
requests==2.32.3
